+++ frontend/src/App.css
@@ -59,31 +59,75 @@ button{
   border: none;
 }
 
-.contentContainer{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-left:200px; 
+.contentContainer {
+  display: grid;
+  grid-template-rows: auto 1fr; 
+  align-content: start;
+  margin: 200px; 
   width: 100%; 
-  min-height: 80vh;
   background-color: rgb(240, 240, 240);
+}
+
+.tabContainer{
+  display:flex;
+  width: 20%;
+  background-color: #ccc;
+}
+.tab {
+  width: 100%;
+  height: 50px;
+  background-color: #ccc; 
+  color: black;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: bold;
 }
 
 .content{
   display: flex;
   flex-direction: center;
+  align-items: center;
+  justify-content: center;
   width: 100%;
-  height: 75%;
+  height: 100%;
   background-color: rgb(255, 255, 255);
 }
 
+
 .form{
   border: 1px solid black;
-  width: 50%;
-  height: 75%;
+  border-radius: 10%;
+  height: 30%;
+  padding: 100px;
+  margin: 30px;
 }
 
+.tableContent {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0; 
+}
+.tableContent h2{
+  margin-bottom: 0;
+  margin-left:30px;
+}
 
+table{
+  margin: 30px;
+  border-collapse: collapse;
+  background-size: 50px 50px;
+  width: 500px; 
+  height: 300px; 
+  table-layout: fixed; 
+}
+td, th {
+  border: 1px solid black;
+  width: 50px;
+  height: 50px;
+  text-align: center;
+}
 @media (max-width: 768px) {
   .sidebar {
     width: 60px;
