+++ frontend/src/pages/Inventory/inventory.jsx
@@ -1,0 +1,9 @@
+export default function Inventory(){
+    return (
+        <div className="contentContainer">
+            <div className="content">
+                <h1>This should have the inventory content</h1>
+            </div>
+        </div>
+    )
+}
