+++ frontend/src/pages/Records/records.jsx
@@ -1,0 +1,13 @@
+import '../../App.css'
+
+export default function Records(){
+    return (
+        <>  
+            <div>
+                <div className="content">
+                    <h1>This should have the records content</h1>
+                </div>
+            </div>
+        </>   
+    )
+}
