--- frontend/src/pages/inventory.jsx
@@ -1,3 +1,0 @@
-export default function Inventory(){
-    return <h1>This is the Inventory tab</h1>
-}
