--- frontend/src/pages/records.jsx
@@ -1,3 +1,0 @@
-export default function Records(){
-    return <h1>This is the Records tab</h1>
-}
